[EI-133] [web] refactor: create atom folder and change path
