+++ web/app/ui/components/numerical-guidance/atom/metadata-list-item.tsx
@@ -1,9 +1,9 @@
-import SelectableListItem from '../view/atom/selectable-list-item';
-import IconButton from '../view/atom/icon-button/icon-button';
+import SelectableListItem from '../../view/atom/selectable-list-item';
+import IconButton from '../../view/atom/icon-button/icon-button';
 import { DotsHorizontalIcon } from '@heroicons/react/solid';
 import { IndicatorBoardMetadataResponse } from '@/app/store/querys/numerical-guidance/indicator-board-metadata.query';
 import { useSelectedIndicatorBoardMetadata } from '@/app/business/hooks/use-selected-indicator-board-metadata.hook';
-import { useDialogMenu } from '../view/molocule/dialog-menu';
+import { useDialogMenu } from '../../view/molocule/dialog-menu';
 import { DIALOG_KEY } from '@/app/utils/keys/dialog-key';
 
 type MetadataListItemProps = {
